[refactor] Member 엔티티 패키지 내에 Type 패키지 추가
설명:
- Type 관련 클래스들을 모아둘 패키지를 추가하려 합니다.
- RoleType 엔티티를 Type 패키지 하위로 이동했습니다.
