+++ src/main/java/com/coolpeace/domain/member/repository/RoleRepository.java
@@ -1,7 +1,7 @@
 package com.coolpeace.domain.member.repository;
 
 import com.coolpeace.domain.member.entity.Role;
-import com.coolpeace.domain.member.entity.RoleType;
+import com.coolpeace.domain.member.entity.type.RoleType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
